Move AsyncLock to Client
