+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/Internal/Threading/AsyncLock.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 46f8d33b46dba1a4aaf1997c6219a34a
+guid: 080606800be7cc4459aac728acf65083
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
